+++ cognite/neat/workflows/steps/lib/rules_importer.py
@@ -3,7 +3,7 @@ import logging
 import time
 from datetime import datetime
 from pathlib import Path
-from typing import ClassVar
+from typing import ClassVar, cast
 
 import yaml
 from rdflib import Namespace
@@ -17,7 +17,55 @@ from cognite.neat.workflows.steps.step_model import Configurable, Step
 
 CATEGORY = __name__.split(".")[-1].replace("_", " ").title()
 
-__all__ = ["OpenApiToRules", "ArbitraryJsonYamlToRules", "GraphToRules"]
+__all__ = ["OpenApiToRules", "ArbitraryJsonYamlToRules", "GraphToRules", "OntologyToRules"]
+
+
+class OntologyToRules(Step):
+    """The step extracts schema from OpenApi/Swagger specification and generates NEAT transformation rules object."""
+
+    description = "The step extracts NEAT rules object from OWL Ontology and \
+        exports them as an Excel rules files for further editing."
+    category = CATEGORY
+    version = "0.1.0-alpha"
+    configurables: ClassVar[list[Configurable]] = [
+        Configurable(
+            name="ontology_file_path",
+            value="staging/ontology.ttl",
+            label="Relative path to the OWL ontology file.",
+        ),
+        Configurable(
+            name="excel_file_path",
+            value="staging/rules.xlsx",
+            label="Relative path for the Excel rules storage.",
+        ),
+    ]
+
+    def run(self) -> FlowMessage:  # type: ignore[override, syntax]
+        ontology_file_path = self.data_store_path / Path(self.configs["ontology_file_path"])
+        excel_file_path = self.data_store_path / Path(self.configs["excel_file_path"])
+        report_file_path = excel_file_path.parent / f"report_{excel_file_path.stem}.txt"
+
+        rules = importer.OWLImporter(ontology_file_path).to_rules(skip_validation=True)
+        exporter.ExcelExporter(rules=cast(Rules, rules), filepath=excel_file_path).export()
+
+        if report := importer.ExcelImporter(filepath=excel_file_path).to_raw_rules().validate_rules():
+            report_file_path.write_text(report)
+
+        relative_excel_file_path = str(excel_file_path).split("/data/")[1]
+        relative_report_file_path = str(report_file_path).split("/data/")[1]
+
+        output_text = (
+            "<p></p>"
+            "Rules generated from OWL Ontology can be downloaded here : "
+            f'<a href="http://localhost:8000/data/{relative_excel_file_path}?{time.time()}" '
+            f'target="_blank">{excel_file_path.stem}.xlsx</a>'
+            "<p></p>"
+            "Report can be downloaded here : "
+            f'<a href="http://localhost:8000/data/{relative_report_file_path}?{time.time()}" '
+            f'target="_blank">{report_file_path.stem}.txt</a>'
+        )
+
+        return FlowMessage(output_text=output_text)
 
 
 class OpenApiToRules(Step):